third_party/blink/web_tests/external/wpt/css/css-flexbox/abspos/flex-abspos-staticpos-align-self-rtl-004-expected.txt

This is a testharness.js-based test.
PASS .container > div 1
PASS .container > div 2
PASS .container > div 3
PASS .container > div 4
FAIL .container > div 5 assert_equals: 
<div class="container"><div style="align-self: last baseline" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 10
PASS .container > div 6
PASS .container > div 7
FAIL .container > div 8 assert_equals: 
<div class="container"><div style="align-self: end" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 10
FAIL .container > div 9 assert_equals: 
<div class="container"><div style="align-self: self-start" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 10
PASS .container > div 10
PASS .container > div 11
PASS .container > div 12
PASS .container > div 13
PASS .container > div 14
PASS .container > div 15
PASS .container > div 16
FAIL .container > div 17 assert_equals: 
<div class="container"><div style="align-self: last baseline" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got -2
PASS .container > div 18
PASS .container > div 19
FAIL .container > div 20 assert_equals: 
<div class="container"><div style="align-self: end" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got -2
FAIL .container > div 21 assert_equals: 
<div class="container"><div style="align-self: self-start" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got -2
PASS .container > div 22
PASS .container > div 23
PASS .container > div 24
Harness: the test ran to completion.
